feat: isolate and track forester worker concurrency

+++ prover/server/prover/v1/non_inclusion_test.go
@@ -226,7 +226,7 @@ func TestNonInclusionCircuit(t *testing.T) {
 			LeafLowerRangeValue:  *leafLowerRangeValue,
 			LeafHigherRangeValue: *leafHigherRangeValue,
 			NextIndex:            uint32(0),
-			PathIndex:            uint32(pathIndex),
+			PathIndex:            uint64(pathIndex),
 			PathElements:         pathElements,
 		}
